// app/containers/RoleManage/types.ts
export interface IDavinciUser {
  id: number
  username: string
  email: string
  role: number
  avatar: string
}
export interface IOrganization {
  id: number
  name: string
  role?: number
  avatar: string
  roleNum?: number
  memberNum?: number
  projectNum?: number
  description: string
  memberPermission?: number
  allowCreateProject?: number
  allowChangeVisibility?: number
  allowDeleteOrTransferProject?: number
}

export interface IOrganizationMember {
  id: number
  user: IDavinciUser
}

export interface IOrganizationRole {
  id: number
  name: string
  description: string
}
// ...
type onGetRoleListByMemberId = (orgId: number, memberId: number, resolve?: (res: any) => void) => void
// ...
export interface IOrganizationProps {
  loginUser: any
  organizations: any
  starUserList: IStarUser[]
  inviteMemberList: any
  currentOrganization: IOrganization
  collectProjects: IProject[]
  currentOrganizationRole: IOrganizationRole[]
  onLoadOrganizationProjects: (param: {id: number, pageNum?: number, pageSize?: number}) => any
  onLoadOrganizationMembers: (id: number) => any
  onLoadOrganizationDetail: (id: number) => any
  onDeleteOrganizationMember: (id: number, resolve: () => any) => any
  onChangeOrganizationMemberRole: (id: number, role: number, resolve: () => any) => any
  currentOrganizationProjects: IProject[]
  currentOrganizationProjectsDetail: {total?: number, list: IProject[]}
  currentOrganizationMembers: IOrganizationMember[]
  onInviteMember: (ordId: number, members: number[], needEmail: boolean, resolve: () => void) => any
  onSearchMember: (keywords: string) => any
  onClickCollectProjects: (formType: string, project: object, resolve: (id: number) => any) => any
  onLoadCollectProjects: () => any
  onEditOrganization: (organization: IOrganization) => any
  onDeleteOrganization: (id: number, resolve: () => any) => any
  onCheckUniqueName: (pathname: any, data: any, resolve: () => any, reject: (error: string) => any) => any
  onGetRoleListByMemberId: onGetRoleListByMemberId
  onLoadOrganizationRole: (id: number) => any
}